feat(filterrrules): update contants file and style
