+++ src/filterRules/demos/basic.tsx
@@ -2,21 +2,8 @@ import React, { useState } from 'react';
 import { Button, Input } from 'antd';
 import { FilterRules } from 'dt-react-component';
 import { IComponentProps } from 'dt-react-component/filterRules';
-import shortid from 'shortid';
 
-export const INIT_ROW_VALUES = {
-    input: '',
-};
-
-export const INIT_DATA = {
-    key: shortid(),
-    level: 0,
-    rowValues: {
-        ...INIT_ROW_VALUES,
-    },
-};
-
-export type IRow = typeof INIT_ROW_VALUES;
+import { INIT_DATA, INIT_ROW_VALUES, IRow } from './constants';
 
 const MyInput = ({ rowValues: { input }, rowKey, onChange }: IComponentProps<IRow>) => (
     <Input value={input} onChange={(e) => onChange?.(rowKey, { input: e.target.value })} />
